advanced_rag_agent/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
chromadb
tiktoken
pypdf
docx2txt
python-dotenv
rank_bm25
flashrank
